+++ PolyvVodSDKDemo/PolyvVodSDKDemo/Classes/Download/Model/PLVDownloadCompleteInfoModel.m
@@ -1,0 +1,14 @@
+//
+//  PLVDownloadCompleteInfoModel.m
+//  PolyvVodSDKDemo
+//
+//  Created by mac on 2018/8/18.
+//  Copyright © 2018年 POLYV. All rights reserved.
+//
+
+#import "PLVDownloadCompleteInfoModel.h"
+#import <PLVVodSDK/PLVVodSDK.h>
+
+@implementation PLVDownloadCompleteInfoModel
+
+@end